// src/types/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from "contentful"
import { Document } from "@contentful/rich-text-types"

export interface ISocialMedia {
  href: string
  name: string
  value: string
}

export interface IAuthorFields {
  /** name */
  name?: string | undefined

  /** resume */
  resume?: string | undefined

  /** socialMedia */
  socialMedia?: ISocialMedia[] | undefined

  /** Profile Pic */
  profilePic?: Asset | undefined

  /** bio */
  bio?: string | undefined

  /** services */
  services?: IService[] | undefined
}

export interface IAuthor extends Entry<IAuthorFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: "author"
        linkType: "ContentType"
        type: "Link"
      }
    }
  }
}

export interface IBlogPostFields {
  /** title */
  title: string

  /** slug */
  slug?: string | undefined

  /** Publish Date */
  publishDate?: string | undefined

  /** resume */
  resume?: string | undefined

  /** Body */
  body?: Document | undefined

  /** tags */
  tags?: Entry<{ [fieldId: string]: unknown }>[] | undefined

  /** Featured Image */
  featuredImage?: Asset | undefined
}

/** Blog post entry */

export interface IBlogPost extends Entry<IBlogPostFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: "blogPost"
        linkType: "ContentType"
        type: "Link"
      }
    }
  }
}

export interface IProjectFields {
  /** title */
  title?: string | undefined

  /** description */
  description?: Document | undefined

  /** date */
  date?: string | undefined

  /** preview */
  preview?: Asset | undefined

  /** url */
  url?: string | undefined
}

/** projects in which he contributed */

export interface IProject extends Entry<IProjectFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: "project"
        linkType: "ContentType"
        type: "Link"
      }
    }
  }
}

export interface IServiceFields {
  /** Title */
  title?: string | undefined

  /** Description */
  description?: Document | undefined

  /** featuredImage */
  featuredImage?: Asset | undefined

  /** active */
  active?: boolean | undefined

  /** Examples */
  examples?: IProject[] | undefined
}

/** descripción de los servicios que ofresco */

export interface IService extends Entry<IServiceFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: "service"
        linkType: "ContentType"
        type: "Link"
      }
    }
  }
}

export interface ITagFields {
  /** Tag */
  tag: string
}

/** Blog post tag */

export interface ITag extends Entry<ITagFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: "tag"
        linkType: "ContentType"
        type: "Link"
      }
    }
  }
}

export type CONTENT_TYPE = "author" | "blogPost" | "project" | "service" | "tag"

export type LOCALE_CODE = "es-MX"

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "es-MX"